[refactor] Change DIR DDD 유지 related to #36
