+++ search/src/main/java/com/bjcareer/search/application/search/RankingService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import com.bjcareer.search.application.port.in.RankingUsecase;
-import com.bjcareer.search.repository.cache.CacheRepository;
+import com.bjcareer.search.out.repository.cache.CacheRepository;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
